@@ -1,22 +1,25 @@
-import React, { Component } from 'react';
-import { Menu, Sider } from 'antd';
-import HomePage from './HomePage';
-import './App.css';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import React from "react";
+import LoginPage from "./pages/Login"
+import HomePage from "./HomePage";
+import UserManagement from "./pages/UserManagement";
+import "./App.css";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { ProtectedRoute } from "./pages/ProtectedRoute";
+import { AuthProvider } from "./hooks/useAuth";
 
 function App() {
   return (
-    <Router>      
-      <Sider width={200} theme="light">
-        <Menu mode="vertical" theme="light" defaultSelectedKeys={['1']}>
-          <Menu.Item key="1" icon={<FileAddOutlined />}>
-            Transaction Processing
-          </Menu.Item>
-        </Menu>
-      </Sider>
-      <Route exact path="/" component={HomePage} />
-      {/* Add more routes for other pages if needed */}
-    </Router>
+
+    <BrowserRouter>
+      <AuthProvider>
+        <Routes>
+          <Route path="login" element={<LoginPage/>}/>
+          <Route path="/" element={<ProtectedRoute><HomePage/></ProtectedRoute>}/>
+          <Route path="usersManagement" element={<ProtectedRoute><UserManagement/></ProtectedRoute>}/>
+        </Routes>
+      </AuthProvider>
+    </BrowserRouter>
+
   );
 }
 
