@@ -1,14 +1,9 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import HomePage from './HomePage'; // Import your homepage component
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "./App"
+
 
 ReactDOM.render(
-  <Router>
-    <Routes>
-      <Route path="/" element={<HomePage/>}/>
-        {/* Add more routes for other pages */}
-    </Routes>
-  </Router>,
-  document.getElementById('root')
+  <App/>,
+  document.getElementById("root")
 );
