@@ -1,12 +1,12 @@
-'use strict';
+"use strict";
 
-const fs = require('fs');
-const path = require('path');
-const Sequelize = require('sequelize');
-const process = require('process');
+const fs = require("fs");
+const path = require("path");
+const Sequelize = require("sequelize");
+const process = require("process");
 const basename = path.basename(__filename);
-const env = process.env.NODE_ENV || 'development';
-const config = require(__dirname + '/../config/config.json')[env];
+const env = process.env.NODE_ENV || "development";
+const config = require(__dirname + "/../config/config.json")[env];
 const db = {};
 
 // Initialize Sequelize
@@ -21,10 +21,10 @@ fs
   .readdirSync(__dirname)
   .filter(file => {
     return (
-      file.indexOf('.') !== 0 &&
+      file.indexOf(".") !== 0 &&
       file !== basename &&
-      file.slice(-3) === '.js' &&
-      file.indexOf('.test.js') === -1
+      file.slice(-3) === ".js" &&
+      file.indexOf(".test.js") === -1
     );
   })
   .forEach(file => {
@@ -42,8 +42,11 @@ db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
 // Import all the models
-const TransactionModel = require('./Transaction'); 
+const TransactionModel = require("./Transaction");
+const UserModal = require("./User")
 const Transaction = TransactionModel(sequelize, Sequelize);
+const User = UserModal(sequelize);
 db.Transaction = Transaction;
+db.User = User;
 
 module.exports = db;
